# Speed up the loading of the block index