--- a/src/test/checkblock_tests.cpp
+++ b/src/test/checkblock_tests.cpp
@@ -44,7 +44,8 @@ bool read_block(const std::string &filename, CBlock &block)
 bool LockAndContextualCheckBlock(CBlock &block, CValidationState &state)
 {
     LOCK(cs_main);
-    return ContextualCheckBlock(block, state, nullptr, false);
+    CCoinsViewCache view(pcoinsTip);
+    return ContextualCheckBlock(block, view, state, nullptr, false);
 }
 
 BOOST_FIXTURE_TEST_SUITE(checkblock_tests, BasicTestingSetup) // BU harmonize suite name with filename