--- a/src/validation/validation.h
+++ b/src/validation/validation.h
@@ -101,6 +101,7 @@ void InvalidChainFound(CBlockIndex *pindexNew);
 
 /** Context-dependent validity block checks */
 bool ContextualCheckBlock(const CBlock &block,
+    CCoinsViewCache &view,
     CValidationState &state,
     CBlockIndex *pindexPrev,
     const bool fConservative = false);