When an operator reboots their computer and starts up the node
it can take much longer to startup because the block index files
are not already in the OS cache. Startup times are improved by roughly
300% by first caching the leveldb files in memory. To effect this we
read the contents of each leveldb file before actually using leveldb
to iterate through the blockindex entries in those files. This way when
leveldb tries to access each entry it doesn't have to go to the disk
each time but rather can read directly from the OS file cache.

Testing:
1) To test this out you must first reboot your computer to clear the OS cache
2) Launch the release or dev branch version and wait for startup to complete
3) Go to the log file and find the time it took to load the index by viewing the two following
entries that define when we started and finished the loading of the block index and manually
calculate the time differential.
2020-01-29 15:41:19.141754 init message: Loading block index...
2020-01-29 15:41:44.571349 LoadBlockIndexDB: hashBestChain=000000000000000000c893b587f9509c947246d0d194df2e64d9f96846723721 height=619977 date=2020-01-29 15:30:31 progress=0.999963

4) Do the above 3 steps again using a build from this PR and note the difference.


My own results for mainnet are as follows:
1) Current dev branch:  25 seconds
2) This PR:  8 seconds

My own results for testnet are as follows:
1) Current dev branch:  49 seconds
2) This PR:  15 seconds